Cambios:
- Implementados todos los casos de uso para el area de ventas
- Modificacion de application.yaml para soporte de Flyway
- Creacion de schema de la DB y modificacion de la tabla rol

+++ Backend/muebleria/src/main/java/com/erp/muebleria/modules/administracion/application/dto/ActualizarConfiguracionDTO.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class ActualizarConfiguracionDTO {
 
-    private Double tasaIva;
+    private BigDecimal tasaIva;
     private String metodoValoracion;
     private String resolucionFactura;
     private String serieFacturas;
